lib: refactor authCallback
Make it into ES6 class, rename to SimpleAuthPolicy
