+++ jstp.js
@@ -17,4 +17,4 @@ jstp.tcp = require('./lib/transport.tcp');
 jstp.ipc = require('./lib/transport.ipc');
 jstp.ws = require('./lib/transport.ws');
 
-jstp.simpleAuth = require('./lib/simple-auth');
+jstp.SimpleAuthPolicy = require('./lib/simple-auth-policy');
