+++ lib/server.js
@@ -6,7 +6,7 @@ const util = require('util');
 const apps = require('./applications');
 const common = require('./common');
 const Connection = require('./connection');
-const simpleAuth = require('./simple-auth');
+const SimpleAuthPolicy = require('./simple-auth-policy');
 
 const HANDSHAKE_TIMEOUT = 3000;
 
@@ -15,26 +15,29 @@ module.exports = Server;
 // JSTP server
 //   rawServer - an underlying server or its adapter
 //   applications - applications index
-//   authCallback - authentication callback (optional)
+//   authPolicy - authentication policy (optional)
+//                see jstp.SimpleAuthPolicy
 //   heartbeatInterval - heartbeat interval, if heartbeat should be used
 //                       (optional)
 //
-function Server(rawServer, applications, authCallback, heartbeatInterval) {
+function Server(rawServer, applications, authPolicy, heartbeatInterval) {
   events.EventEmitter.call(this);
 
   if (Array.isArray(applications)) {
     applications = apps.createAppsIndex(applications);
   }
 
-  if (typeof(authCallback) === 'number') {
-    heartbeatInterval = authCallback;
-    authCallback = null;
+  if (typeof(authPolicy) === 'number') {
+    heartbeatInterval = authPolicy;
+    authPolicy = null;
   }
 
   this.rawServer = rawServer;
   this.applications = applications;
-  this.startSession = authCallback || simpleAuth.startSession;
   this.heartbeatInterval = heartbeatInterval;
+
+  if (!authPolicy) authPolicy = new SimpleAuthPolicy();
+  this.startSession = (...args) => authPolicy.startSession(...args);
 
   this.clients = {};
   this._cachedClientsArray = [];
